+++ tests/specs/node-client.test.js
@@ -2,13 +2,14 @@ import { Buffer } from 'buffer'
 import FormData from 'form-data'
 
 import Request from '../../src'
+import { cache } from '../../src/cache'
 
 import { registerNodeTransaction, wait } from '../helpers'
 
 describe('Node Client', () => {
 
   afterEach(() => {
-    Request.resetAllCache()
+    cache.deleteAll()
   })
 
   describe('Request Options', () => {
@@ -743,8 +744,8 @@ describe('Node Client', () => {
     })
 
     it('resets all caches by ttl', async () => {
-      const oldFlushInterval = Request.__cache.flushInterval
-      Request.__cache.setFlushInterval(500)
+      const oldFlushInterval = cache.flushInterval
+      cache.setFlushInterval(500)
 
       registerNodeTransaction(['result1'])
       registerNodeTransaction(['result2'])
@@ -762,7 +763,7 @@ describe('Node Client', () => {
       expect(result1).toBe('result1')
       expect(result2).toBe('result2')
 
-      Request.__cache.setFlushInterval(oldFlushInterval)
+      cache.setFlushInterval(oldFlushInterval)
     })
 
     it('resets cache by POST request', async () => {
